+++ server/src/main/java/com/codestates/seb42_pre_031/answer/dto/AnswerPatchDto.java
@@ -11,7 +11,7 @@ public class AnswerPatchDto {
     @NotBlank
     private String contents;
 
-    private int answerId;
+    private long answerId;
 
-    private int questionId;
+    private long questionId;
 }
