+++ server/src/main/java/com/codestates/seb42_pre_031/answer/dto/AnswerResponseDto.java
@@ -7,9 +7,9 @@ import lombok.Setter;
 @Getter
 @AllArgsConstructor
 public class AnswerResponseDto {
-    private int answerId;
-    private int questionId;
-    private int memberId;
+    private long answerId;
+    private long questionId;
+    private long memberId;
     private String contents;
    // private int voteCount;
 }
