+++ server/src/main/java/com/codestates/seb42_pre_031/answer/repository/AnswerRepository.java
@@ -3,5 +3,5 @@ package com.codestates.seb42_pre_031.answer.repository;
 import com.codestates.seb42_pre_031.answer.entity.Answer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AnswerRepository extends JpaRepository<Answer, Integer> {
+public interface AnswerRepository extends JpaRepository<Answer, long> {
 }
